+++ fastfold/model/fastnn/kernel/jit/fused_ops.py
@@ -7,7 +7,7 @@ def bias_sigmod_ele(y, bias, z):
     return torch.sigmoid(y + bias) * z
 
 
-@torch.jit.script
+# @torch.jit.script
 def bias_dropout_add(x: torch.Tensor, bias: torch.Tensor, dropmask: torch.Tensor,
                      residual: torch.Tensor, prob: float, training: bool) -> torch.Tensor:
     out = (x + bias) * F.dropout(dropmask, p=prob, training=training)
