+++ fastfold/model/fastnn/msa.py
@@ -3,7 +3,8 @@ import math
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from fastfold.model.fastnn.kernel import LayerNorm
+# from fastfold.model.fastnn.kernel import LayerNorm
+from torch.nn import LayerNorm
 
 from fastfold.model.fastnn.ops import Transition, SelfAttention
 from fastfold.model.fastnn.kernel import bias_dropout_add
